+++ crates/percore/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.2.3
+
+### Bugfixes
+
+- Fixed broken build on aarch32.
+
+## 0.2.2
+
+### New features
+
+- Added `PerCore::get_mut`.
+- Added support for AArch32.
+
 ## 0.2.1
 
 ### New features